+++ hkbbs-web/src/utils/api/index.ts
@@ -1,9 +1,7 @@
 import { Card } from "./helper/card"
-import { Hole } from "./helper/hole"
 import { Post } from "./helper/post"
-import { Both } from "./helper/both"
 import { Notice } from "./helper/notice"
 import { User } from "./helper/user"
 import { Chat } from "./helper/chat"
 
-export default { Card, Hole, Post, Both, Notice, User, Chat }
+export default { Card, Post, Notice, User, Chat }
